requirements.txt
gTTS==2.5.4
Pillow==11.1.0
